# docs/spec/api-contracts.yaml
openapi: 3.0.3
info:
  title: ContestForge Core Contest Engine API
  version: 1.0.0
  description: >
    Multi-tenant live contest engine. REST endpoints cover tenancy, user
    management, contest authoring, lifecycle, registration, live-runtime
    controls, results, notifications, and audit. Live question delivery,
    answer submission, leaderboard, and elimination events flow over a
    WebSocket channel (documented in api-contracts.md §WebSocket).

    Authentication: every endpoint except /auth/login, /auth/refresh, /health,
    and /ready requires a JWT bearer token whose claims include the caller's
    role and tenant_id. Cross-tenant access is rejected with 403.

    Onboarding: the first Super Admin is bootstrapped by an SQL script and is
    not exposed in this API. A Super Admin creates an organization (which
    provisions its initial Org Admin). An Org Admin creates co-Org-Admins,
    Moderators, and Participants via POST /users. There is no public self-signup.
servers:
  - url: https://api.contestforge.example/v1
    description: Placeholder; region/host finalized in /neutron:plan
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and obtain JWT access/refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new token pair (rotates the token)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the presented refresh token (and its rotation chain)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '204': { description: Logged out (token revoked) }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [Auth]
      summary: Return the current authenticated principal
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, format: password, minLength: 8 }
      responses:
        '204': { description: Password changed }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ───────────────────────────────── Users ─────────────────────────────────
...
  /contests:
    post:
      tags: [Contests]
      summary: Create a contest (Draft). Org Admin.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateContestRequest' }
      responses:
        '201':
          description: Contest created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    get:
      tags: [Contests]
      summary: List contests in the caller's tenant
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/LifecycleStatus' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: Contests
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Contest' }
  /contests/{contestId}:
    parameters:
      - { $ref: '#/components/parameters/ContestId' }
    get:
      tags: [Contests]
      summary: Get a contest (with groups + configuration)
      responses:
        '200':
          description: Contest
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContestDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Contests]
      summary: Update contest metadata. Allowed only while Draft.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateContestRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contest' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Contests]
      summary: Delete a contest. Allowed only while Draft.
      responses:
        '204': { description: Deleted }
        '409': { $ref: '#/components/responses/Conflict' }
  /contests/{contestId}/lifecycle:
    post:
      tags: [Lifecycle]
      summary: Transition the contest to the next lifecycle stage
      description: >
        Advances along the fixed sequence DRAFT → PUBLISHED → REGISTRATION_OPEN
        → REGISTRATION_CLOSED → SCHEDULED → LIVE → COMPLETED → ARCHIVED. Skipping
        a stage or unmet preconditions return 409. SCHEDULED requires
        scheduled_start_at.
      parameters:
        - { $ref: '#/components/parameters/ContestId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_status]
              properties:
                target_status: { $ref: '#/components/schemas/LifecycleStatus' }
                scheduled_start_at:
                  type: string
                  format: date-time
                  description: Required when transitioning to SCHEDULED
      responses:
        '200':
          description: Transitioned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contest' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ───────────────────────────────── Groups ────────────────────────────────
  /contests/{contestId}/groups:
    parameters:
      - { $ref: '#/components/parameters/ContestId' }
    post:
      tags: [Groups]
      summary: Add a group to a Grouped contest (Draft only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateGroupRequest' }
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Groups]
      summary: List groups for a Grouped contest
      responses:
        '200':
          description: Groups
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Group' }
  /contests/{contestId}/groups/{groupId}:
    parameters:
      - { $ref: '#/components/parameters/ContestId' }
      - { $ref: '#/components/parameters/GroupId' }
    patch:
      tags: [Groups]
      summary: Update a group (name, sequence, weight). Draft only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                sequence: { type: integer }
                weight: { type: number, nullable: true }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Groups]
      summary: Delete a group. Draft only.
      responses:
        '204': { description: Deleted }
        '409': { $ref: '#/components/responses/Conflict' }

  # ───────────────────────────── Configuration ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ContestId:
      in: path
      name: contestId
      required: true
      schema: { type: string, format: uuid }
    GroupId:
      in: path
      name: groupId
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
      description: Opaque cursor returned by the previous page
  responses:
    BadRequest:
      description: Malformed request (schema/parse error)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Role or tenant not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found in caller's tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Invalid state transition or precondition failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UnprocessableEntity:
      description: Semantic validation error (e.g. elimination block missing rules)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
    Page:
      type: object
      description: Pagination envelope; concrete `items` defined per endpoint
      properties:
        next_cursor: { type: string, nullable: true }
        has_more: { type: boolean }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        tenant_slug:
          type: string
          description: Tenant hint for non-super-admin users (matches Organization.slug)
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: bearer }
        expires_in: { type: integer, example: 900 }
        role: { type: string, enum: [SUPER_ADMIN, ORG_ADMIN, MODERATOR, PARTICIPANT] }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid, nullable: true }
        email: { type: string, format: email }
        first_name: { type: string }
        last_name: { type: string }
        role: { type: string, enum: [SUPER_ADMIN, ORG_ADMIN, MODERATOR, PARTICIPANT] }
        status: { type: string, enum: [ACTIVE, DISABLED] }
        created_at: { type: string, format: date-time }
...
    LifecycleStatus:
      type: string
      enum:
        - DRAFT
        - PUBLISHED
        - REGISTRATION_OPEN
        - REGISTRATION_CLOSED
        - SCHEDULED
        - LIVE
        - COMPLETED
        - ARCHIVED
    Contest:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        structure: { type: string, enum: [NORMAL, GROUPED] }
        lifecycle_status: { $ref: '#/components/schemas/LifecycleStatus' }
        scheduled_start_at: { type: string, format: date-time, nullable: true }
        group_score_rollup:
          type: string
          enum: [SUM, WEIGHTED_SUM, BEST_N]
          nullable: true
        rollup_best_n: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
    ContestDetail:
      allOf:
        - $ref: '#/components/schemas/Contest'
        - type: object
          properties:
            groups:
              type: array
              items: { $ref: '#/components/schemas/Group' }
            configuration:
              type: array
              items: { $ref: '#/components/schemas/ConfigurationBlock' }
    CreateContestRequest:
      type: object
      required: [name, structure]
      properties:
        name: { type: string }
        description: { type: string }
        structure: { type: string, enum: [NORMAL, GROUPED] }
        group_score_rollup: { type: string, enum: [SUM, WEIGHTED_SUM, BEST_N] }
        rollup_best_n: { type: integer }
    UpdateContestRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
    CreateGroupRequest:
      type: object
      required: [name, sequence]
      properties:
        name: { type: string }
        sequence: { type: integer }
        weight: { type: number, nullable: true }
    Group:
      type: object
      properties:
        id: { type: string, format: uuid }
        contest_id: { type: string, format: uuid }
        name: { type: string }
        sequence: { type: integer }
        weight: { type: number, nullable: true }
    ConfigurationBlock:
      type: object
      required: [mode, question_duration_s]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        group_id:
          type: string
          format: uuid
          nullable: true
          description: Required for Grouped contests; null/omitted for Normal
        mode: { type: string, enum: [STANDARD, SPEED, ELIMINATION] }
        question_duration_s: { type: integer, minimum: 5, maximum: 300 }
        question_interval_s: { type: integer, minimum: 0, maximum: 60 }
        explanation_duration_s: { type: integer, minimum: 0, maximum: 60 }
        leaderboard_duration_s: { type: integer, minimum: 0, maximum: 60 }
        reveal_mode:
          type: string
          enum: [AUTOMATIC, MODERATOR_CONTROLLED]
        ranking_criterion:
          type: string
          enum: [SCORE_ONLY, SCORE_TIME, ACCURACY]
        tie_display:
          type: string
          enum: [SHARED_RANK, FASTEST, LEAST_INCORRECT]
        leaderboard_visibility:
          type: string
          enum: [ALWAYS, POST_QUESTION, HIDDEN, MASKED]
        update_frequency:
          type: string
          enum: [PER_ANSWER, PER_QUESTION, PER_GROUP]
        correct_points: { type: integer, default: 10 }
        second_chance_rate: { type: number, default: 0.5 }
        time_bands:
          type: array
          description: Speed mode scoring bands (overrides defaults)
          items:
            type: object
            properties:
              max_seconds: { type: number }
              points: { type: integer }
        decay:
          type: object
          nullable: true
          description: Optional linear decay alternative to bands (Speed)
          properties:
            max_points: { type: integer }
            floor: { type: integer }
            decay_rate: { type: number }
        survivor_score_reset:
          type: boolean
          default: false
          description: When true, survivors' scores reset at the next group start (FR-37)
        elimination_combine_operator:
          type: string
          enum: [AND, OR]
          nullable: true
          description: Block-level operator combining all elimination rules (ELIMINATION only)
        wildcards:
          type: array
          items: { $ref: '#/components/schemas/WildcardConfig' }
        elimination_rules:
          type: array
          description: Required when mode = ELIMINATION
          items: { $ref: '#/components/schemas/EliminationRule' }
        checkpoints:
          type: array
          description: Required when mode = ELIMINATION
          items: { $ref: '#/components/schemas/Checkpoint' }
    WildcardConfig:
      type: object
      required: [type]
      description: >
        Enables a wildcard type in this block. Each enabled wildcard is usable
        once per participant for the whole contest (no usage_limit, cooldown, or
        carry_over). TOP_50_PERCENT is evaluated against the last committed
        leaderboard at the start of the current question.
      properties:
        type: { type: string, enum: [FIFTY_FIFTY, SECOND_CHANCE, SKIP] }
        eligibility:
          type: string
          enum: [ALL, TOP_50_PERCENT]
          default: ALL
    EliminationRule:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [FIRST_WRONG, N_WRONG, BOTTOM_X_PERCENT, MIN_SCORE]
        n_value: { type: integer, nullable: true }
        percent_value: { type: number, nullable: true }
        min_score: { type: integer, nullable: true }
    Checkpoint:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [AFTER_QUESTION, AFTER_GROUP, CUSTOM_MILESTONE]
        question_sequence: { type: integer, nullable: true }
        milestone_at: { type: string, format: date-time, nullable: true }
...